Finally finished all the endpoint ! still some cleanup todo

+++ app/src/main/java/com/example/epsswim/data/model/app/participation/Swimmerevent.kt
@@ -4,5 +4,5 @@ import com.example.epsswim.data.model.app.participation.swimmingtypes.Eventtype
 
 data class Swimmerevent(
     val eventtype: Eventtype,
-    val laptimes: List<Double>
+    val laptimes: List<Long>
 )
